Trying to make TestChunkingUtils a shared library for creating serialized chunk and manifest messages. 📥
